packages/agents/compliance_counsel/requirements.txt
gradient-adk>=0.1.0
gradient>=0.1.0
langgraph>=0.2.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
